myblog/requirements.txt
asgiref==3.7.2
Django==4.2.5
django-taggit==3.0.0
Markdown==3.4.1
psycopg2-binary==2.9.7
sqlparse==0.4.4
